requirements.txt
certifi==2018.1.18
dj-database-url==0.5.0
Django==2.1.5
django-autocomplete-light==3.3.2
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-disqus==0.5
django-import-export==1.2.0
django-js-asset==1.1.0
django-taggit==0.23.0
gunicorn==19.9.0
Pillow==5.1.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyPDF2==1.26.0
pytz==2018.9
PyYAML==3.13
reportlab==3.4.0
simplejson==3.13.2
whitenoise==4.1.2